requirements.txt
requests==2.11.1
pytest==3.4.2
allure-pytest==2.3.2b1
PyHamcrest==1.9.0
jsonschema==2.6.0
